% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{adj_mod_dic}
\alias{adj_mod_dic}
\title{Adjective Modality Norms}
\format{
A tibble of 423 adjectives with 8 attributes
}
\source{
Lynott, D., & Connell, L. (2009). Modality exclusivity norms for 423 object properties. Behavior Research Methods, 41, 558–564. doi:10.3758/BRM.41.2.558
}
\usage{
adj_mod_dic
}
\description{
adjective modality norms
}
\keyword{datasets}
