% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{auth_dic}
\alias{auth_dic}
\title{Authenticity Dictionary}
\format{
A list of 89  words with one attributes
}
\source{
Boyd, R. L. (2017). TAPA: Textual Affective Properties Analyzer (v.1.1.0) [Software]. Available from https://www.ryanboyd.io/software/tapa
}
\usage{
auth_dic
}
\description{
Authenticity norms
}
\keyword{datasets}
