% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{competence_enet_final}
\alias{competence_enet_final}
\title{Competence Model}
\format{
An object of class \code{train} of length 20.
}
\usage{
competence_enet_final
}
\description{
A pre-trained enet regression model to assess competence perceptions
}
\keyword{datasets}
