% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{environmental_words}
\alias{environmental_words}
\title{Environmental words}
\format{
A list of 42  words
}
\source{
Moss, T. W., Renko, M., Block, E., & Meyskens, M. (2018). Funding the story of hybrid ventures: Crowdfunder lending preferences and linguistic hybridity. Journal of Business Venturing, 33(5), 643-659.
}
\usage{
environmental_words
}
\description{
list of words related to an environmental value orientation
}
\keyword{datasets}
