% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{Example Data}
\format{
A dataframe with 40 rows and 11 columns
}
\usage{
example_data
}
\description{
40 random bios from the vignette data. 20 bios were randomly selected from the competence condition and
20 bios were randomly selected from the warmth condition.
}
\keyword{datasets}
