% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{single_words_dic}
\alias{single_words_dic}
\title{Single Word Norms}
\format{
A tibble of 10,758 words with 13 attributes
}
\source{
\url{https://doomlab.shinyapps.io/single_words/}
}
\usage{
single_words_dic
}
\description{
Single word norms
}
\keyword{datasets}
