% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{social_words}
\alias{social_words}
\title{Social Words}
\format{
A list of 66 words
}
\source{
Moss, T. W., Renko, M., Block, E., & Meyskens, M. (2018). Funding the story of hybrid ventures: Crowdfunder lending preferences and linguistic hybridity. Journal of Business Venturing, 33(5), 643-659.
}
\usage{
social_words
}
\description{
Social Value orientation words
}
\keyword{datasets}
