% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vignette_data}
\alias{vignette_data}
\title{Vignette Data}
\format{
A dataframe with 393 rows and 135 columns
}
\usage{
vignette_data
}
\description{
Sample data from a study that can be used to test and explore the package
}
\keyword{datasets}
