% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{W_C_ratings}
\alias{W_C_ratings}
\title{Warmth and Competence Ratings}
\format{
A list of 1711 words with two attributes
}
\usage{
W_C_ratings
}
\description{
list of words rated by judges on warmth and competence
}
\keyword{datasets}
