% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{warmth_enet_final_reduced}
\alias{warmth_enet_final_reduced}
\title{Warmth Model}
\format{
An object of class \code{train} of length 20.
}
\usage{
warmth_enet_final_reduced
}
\description{
A pre-trained enet regression model to assess warmth perceptions
}
\keyword{datasets}
