% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warmthcompetence-package.R
\docType{package}
\name{warmthcompetence-package}
\alias{warmthcompetence}
\alias{warmthcompetence-package}
\title{warmthcompetence: Warmth and Competence Detectors}
\description{
Detects perceptions of warmth and competence in American English self-presentational language. Using pre-trained enet regression models, this package provides a numerical representation of warmth and competence perceptions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bushraguenoun/warmthcompetence}
  \item Report bugs at \url{https://github.com/bushraguenoun/warmthcompetence/issues}
}

}
\author{
\strong{Maintainer}: Bushra Guenoun \email{bushraguenoun@gmail.com} (\href{https://orcid.org/0000-0003-0975-1620}{ORCID})

Authors:
\itemize{
  \item Julian Zlatev (\href{https://orcid.org/0000-0002-9427-9887}{ORCID})
}

Other contributors:
\itemize{
  \item Noah Greifer \email{ngreifer@iq.harvard.edu} (\href{https://orcid.org/0000-0003-3067-7154}{ORCID}) [contributor]
}

}
\keyword{internal}
