% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/words_clean.R
\name{words_clean}
\alias{words_clean}
\title{A function to format text}
\usage{
words_clean(text, ID)
}
\arguments{
\item{text}{character A vector of texts}

\item{ID}{character A vector of IDs}
}
\value{
Tibbles that are used by the main functions of the \pkg{warmthcompetence} package
}
\description{
Contains functions that are used by the main functions of the \pkg{warmthcompetence} package for text processing.
}
\details{
Some features depend on Spacyr which must be installed separately in Python.
}
\keyword{internal}
