% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warmth.R
\name{warmth}
\alias{warmth}
\title{Warmth Detector}
\usage{
warmth(text, ID, metrics = c("scores", "features", "all"))
}
\arguments{
\item{text}{character A vector of texts, each of which will be assessed for warmth.}

\item{ID}{character A vector of IDs that will be used to identify the warmth scores.}

\item{metrics}{character An argument that allows users to decide what metrics to return. Users can return the warmth scores (metrics = "scores"),
the features that underlie the warmth scores (metrics = "features"), or both the warmth scores and the features (metrics = "all).
The default choice is to return the warmth scores.}
}
\value{
The default is to return a data.frame with each row containing the document identifier and the warmth score.
    Users can also customize what is returned through the metrics argument. If metrics = "features", then a dataframe of warmth features will be
    returned where each document is represented by a row. If metrics = "all", then both the warmth scores and features will be returned in a data.frame.
}
\description{
Assesses warmth perceptions in self-presentational natural language.
    This function is one of the main two functions of the  \code{warmthcompetence} package.
    It takes an N-length vector of self-presentational text documents and N-length vector of document IDs and returns a warmth perception score that represents how much warmth
    others attribute the individual who wrote the self-presentational text.
    The function also contains a metrics argument that enables users to also return the raw features used to assess warmth perceptions.
}
\details{
Some features depend Spacyr which must be installed seperately in Python.
}
\examples{

data("vignette_data")

warmth_scores <- warmth(vignette_data$Message, vignette_data$ResponseId, metrics = "all")

vignette_data$warmth_predictions <- warmth_scores$warmth_predictions

warmth_model2 <- lm(RA_warm_AVG  ~ warmth_predictions, data = vignette_data)
summary(warmth_model2)

}
\references{
Yeomans, M., Kantor, A., & Tingley, D. (2018). The politeness Package: Detecting Politeness in Natural Language. R Journal, 10(2).
Rinker, T. W. (2018). lexicon: Lexicon Data version 1.2.1. http://github.com/trinker/lexicon
Moss, T. W., Renko, M., Block, E., & Meyskens, M. (2018). Funding the story of hybrid ventures: Crowdfunder lending preferences and linguistic hybridity. Journal of Business Venturing, 33(5), 643-659.
Buchanan, E. M., Valentine, K. D., & Maxwell, N. P. (2018). LAB: Linguistic Annotated Bibliography - Shiny Application. Retrieved from http://aggieerin.com/shiny/lab_table.
Rinker, T. W. (2013). qdapDictionaries: Dictionaries to Accompany the qdap Package. 1.0.7. University at Buffalo. Buffalo, New York. http://github.com/trinker/qdapDictionaries
Rinker, T. W. (2019). sentimentr: Calculate Text Polarity Sentiment version 2.7.1. http://github.com/trinker/sentimentr
Boyd, R. L. (2017). TAPA: Textual Affective Properties Analyzer (v.1.1.0) [Software]. Available from https://www.ryanboyd.io/software/tapa
}
