% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warmthcompetence-package.R
\docType{package}
\name{warmthcompetence-package}
\alias{warmthcompetence}
\alias{warmthcompetence-package}
\title{warmthcompetence: Warmth and Competence Detectors}
\description{
Detects perceptions of warmth and competence in American English self-presentational language. Using pre-trained enet regression models, this package provides a numerical representation of warmth and competence perceptions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bushraguenoun/warmthcompetence}
  \item Report bugs at \url{https://github.com/bushraguenoun/warmthcompetence/issues}
}

}
\keyword{internal}
